finish

+++ src/app.py
@@ -1,4 +1,17 @@
 import os
 
+
+class FileNotFoundError(Exception):
+    pass
+
+
 def rm(filename):
-    os.remove(filename)
+    if not os.path.isfile(filename):
+        raise FileNotFoundError("The file doesn't exit")
+    else:
+        os.remove(filename)
+        
+
+
+
+
